pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anchor"
version = "7.0.0"
description = "理解模型 — 多模式信息提取与事实验证引擎"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "jinozaki" }]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlmodel>=0.0.21",
    "asyncpg>=0.29.0",
    "alembic>=1.14.0",
    "aiosqlite>=0.20.0",
    # HTTP & Scraping
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    # Social media APIs
    "tweepy>=4.14.0",
    "youtube-transcript-api>=1.2.0",
    "pytubefix>=8.0.0",
    "av>=12.0.0",
    # AI
    "anthropic>=0.40.0",
    # Scheduling
    "apscheduler>=3.10.4",
    # Config
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    # Financial data
    "yfinance>=0.2.40",
    "akshare>=1.14.0",
    # Macro data sources
    "fredapi>=0.5.1",
    "pandas>=2.0.0",
    "pandas-datareader>=0.10.0",
    "wbgapi>=1.0.12",
    "dbnomics>=0.9.3",
    # Web search
    "tavily-python>=0.5.0",
    # Utilities
    "tenacity>=9.0.0",
    "loguru>=0.7.2",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0",
    # CLI
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.4.0",
]

[project.scripts]
anchor = "anchor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["anchor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
